Add Content-Security-Policy
Work and testing was done by @andig so @ypid committed in the name of
andig as andig did all the work and requested a PR :wink:
Also, the git commit is automatically gpg signed by ypid as all commits
committed by him.

Note that ypid has not tested this policy yet.

Closes: #504