don't highlight inactive channels
first, it makes no sense (which is my personal opinion and therefore
debatable). more important, this fixes an itch I always head: due to bad
row indexing, the next active channel's graph got highlighted, which is
definitely wrong.
A minor issue stays: When deactiving a channel (by clicking its checbox)
the row get's highlighted. I will fix this in the next PR by changing
the highlight sensitive area so that the checkbox and the delete/info
actions will not trigger it.