fix php parse errors regarding access of static methods
This fixes several error messages like this:
PHP Parse error:  syntax error, unexpected '::' (T_PAAMAYIM_NEKUDOTAYIM)
in /var/www/volkszaehler.org/lib/Interpreter/SQL/MySQLAggregateOptimizer.php
on line 211

Seen this on a debian stretch installation which contains php7.0.
This are probably not all occasions if this problem, but the ones
resulting in php errors and where the fix was tested.