Zoom for mobile page based on cursor position.
Depend the zoom based on the last hover position. For mobile devices the hover position is set when tap on the chart and is persistant. For normal browsers the hover is done by the mouse and invalid when moving out of the area. Thus the cange only affect mobile devices.